Adding getDatafromCSV.cpp  and myfile.csv
This exercise is from a Bendix interview,
basically they provide a file called myfile.csv,
this file contains persons with birthyear and deadyear,
The goal is to parse this data and find the year with the most people alive.

I created a class Person to store data
and also a class PersonHandler to hanler a person vector
finally a parser to get data from csv file.